# Indexer performance improvements