--- a/Cargo.toml
+++ b/Cargo.toml
@@ -38,7 +38,7 @@ void = "1.0.2"
 xdg = "2.1.0"
 xml-rs = "0.8.0"
 xz2 = "0.1.4"
-zstd = "0.4.17+zstd.1.3.3"
+zstd = "0.6.0+zstd.1.4.8"
 
 [[example]]
 name = "nix-index-debug"