@ncfavier 2023-01-09
This PR does seem to improve the time `nix-index` takes:

```
# before
real	9m13.026s
user	7m36.327s
sys	0m28.164s

# after
real	7m53.152s
user	9m7.771s
sys	0m25.164s
```

@enolan how are you getting 73 seconds? o_O I'm on a reasonably recent CPU (AMD Ryzen 5 4650U).